requirements.txt
pyjwt
tweepy